fix listeners
