+++ web/src/api/db/gameSession.ts
@@ -1,8 +1,6 @@
-import { pipe } from 'fp-ts/lib/pipeable';
 import * as A from 'fp-ts/lib/Array';
-import * as O from 'fp-ts/lib/Option';
-import { extractDocumentData, normalizeDocument } from 'solo-lib/lib/utils/firebase';
-
+import { pipe } from 'fp-ts/lib/pipeable';
+import { normalizeDocument } from 'solo-lib/lib/utils/firebase';
 import { getSessionRef } from '../firebase';
 
 export async function requestGetPlayerHand(
@@ -19,35 +17,6 @@ export async function requestGetPlayerHand(
     normalizedHand,
     A.reduce<Normalized<Card>, Normalized<Card>>({}, (acc, card) => ({ ...acc, ...card })),
   );
-}
-
-export async function requestPlayerHandListener(
-  currentPlayerId: string,
-  sessionId: string,
-  callback: (h: string[]) => void,
-) {
-  try {
-    await getSessionRef(sessionId)
-      .collection('players')
-      .doc(currentPlayerId)
-      .onSnapshot(documentSnapshot => {
-        const player = extractDocumentData<SessionPlayer>(documentSnapshot);
-
-        pipe(
-          player,
-          O.fold(
-            () => {
-              throw new Error('No Player');
-            },
-            p => {
-              callback(p.hand);
-            },
-          ),
-        );
-      });
-  } catch (error) {
-    console.error(error);
-  }
 }
 
 export async function requestAddPlay(sessionId: string, play: Play) {
