+++ web/src/views/Game/index.tsx
@@ -10,6 +10,7 @@ import { useProgressionListener } from './hooks/useProgressionListener';
 import { getPlayerValue } from '../../store/playerHand/selector';
 import { getCurrentCard, getOrderedPlayers } from '../../store/session/selectors';
 import { useTurn } from './hooks/useTurn';
+import { usePlayersListener } from './hooks/usePlayersListener';
 
 export function Game() {
   const players = useSelector(getOrderedPlayers);
@@ -18,6 +19,7 @@ export function Game() {
   const { playerActions } = useHandListener();
   const { hasDrawed, handlePass, handleDrawCard, handleCardClick } = useTurn();
   const { isCurrentPlayer } = useProgressionListener();
+  usePlayersListener();
 
   const onDeckClick = () => {
     if (!hasDrawed) handleDrawCard();
